Extract functions and types from Thermostat capability
Related to #44
