+++ src/hubitat-capabilities/functions/get-supported-thermostat-fan-modes.function.ts
@@ -1,0 +1,35 @@
+/**
+ * @packageDocumentation
+ * @module HubitatCapabilities
+ */
+
+import { HubitatDevice } from '../../hubitat-device-events/hubitat-device';
+import { EThermostatFanModes } from '..';
+import { enumListToStringList, getDevice } from '../capabilities.helpers';
+
+/**
+ * Returns a list of fan modes supported by specified thermostat.
+ *
+ * Capabilities:
+ * - Thermostat
+ *
+ * @param device A target device.
+ * @returns Returns a list of fan modes supported by specified thermostat.
+ */
+export function getSupportedThermostatFanModes(device: HubitatDevice): EThermostatFanModes[];
+
+/**
+ * Returns a list of fan modes supported by specified thermostat.
+ *
+ * Capabilities:
+ * - Thermostat
+ *
+ * @param deviceId An ID of the target device.
+ * @returns Returns a list of fan modes supported by specified thermostat.
+ */
+export function getSupportedThermostatFanModes(deviceId: number): EThermostatFanModes[];
+
+export function getSupportedThermostatFanModes(deviceOrId: HubitatDevice | number): EThermostatFanModes[] {
+  const valueList = getDevice(deviceOrId).getAttributeAsString('supportedThermostatFanModes') ?? '[]';
+  return enumListToStringList(valueList) as EThermostatFanModes[];
+}
