+++ src/hubitat-capabilities/functions/get-thermostat-fan-mode.function.ts
@@ -1,0 +1,42 @@
+/**
+ * @packageDocumentation
+ * @module HubitatCapabilities
+ */
+
+import { HubitatDevice } from '../../hubitat-device-events/hubitat-device';
+import { EThermostatFanModes } from '..';
+import { getDevice } from '../capabilities.helpers';
+
+/**
+ * Returns current thermostat fan mode.
+ *
+ * Capabilities:
+ * - Thermostat
+ * - ThermostatFanMode
+ *
+ * @param device A target device.
+ * @param defaultValue A value returned in case of invalid value returned by Hubitat.
+ * @returns Returns current thermostat fan mode.
+ */
+export function getThermostatFanMode(device: HubitatDevice, defaultValue: EThermostatFanModes): EThermostatFanModes;
+
+/**
+ * Returns current thermostat fan mode.
+ *
+ * Capabilities:
+ * - Thermostat
+ * - ThermostatFanMode
+ *
+ * @param deviceId An ID of the target device.
+ * @param defaultValue A value returned in case of invalid value returned by Hubitat.
+ * @returns Returns current thermostat fan mode.
+ */
+export function getThermostatFanMode(deviceId: number, defaultValue: EThermostatFanModes): EThermostatFanModes;
+
+export function getThermostatFanMode(
+  deviceOrId: HubitatDevice | number,
+  defaultValue: EThermostatFanModes = 'auto',
+): EThermostatFanModes {
+  const fanMode = getDevice(deviceOrId).getAttributeAsString('thermostatFanMode') ?? defaultValue;
+  return fanMode as EThermostatFanModes;
+}
