+++ src/hubitat-capabilities/functions/get-thermostat-mode.function.ts
@@ -1,0 +1,42 @@
+/**
+ * @packageDocumentation
+ * @module HubitatCapabilities
+ */
+
+import { HubitatDevice } from '../../hubitat-device-events/hubitat-device';
+import { EThermostatModes } from '..';
+import { getDevice } from '../capabilities.helpers';
+
+/**
+ * Returns current thermostat mode.
+ *
+ * Capabilities:
+ * - Thermostat
+ * - ThermostatMode
+ *
+ * @param device A target device.
+ * @param defaultValue A value returned in case of invalid value returned by Hubitat.
+ * @returns Returns current thermostat mode.
+ */
+export function getThermostatMode(device: HubitatDevice, defaultValue: EThermostatModes): EThermostatModes;
+
+/**
+ * Returns current thermostat mode.
+ *
+ * Capabilities:
+ * - Thermostat
+ * - ThermostatMode
+ *
+ * @param deviceId An ID of the target device.
+ * @param defaultValue A value returned in case of invalid value returned by Hubitat.
+ * @returns Returns current thermostat mode.
+ */
+export function getThermostatMode(deviceId: number, defaultValue: EThermostatModes): EThermostatModes;
+
+export function getThermostatMode(
+  deviceOrId: HubitatDevice | number,
+  defaultValue: EThermostatModes = 'auto',
+): EThermostatModes {
+  const fanMode = getDevice(deviceOrId).getAttributeAsString('thermostatMode') ?? defaultValue;
+  return fanMode as EThermostatModes;
+}
