+++ src/hubitat-capabilities/index.ts
@@ -1,9 +1,30 @@
 /* FUNCTIONS */
 
 export * from './functions/get-alarm-setting.function';
+export * from './functions/get-cooling-setpoint.function';
+export * from './functions/get-heating-setpoint.function';
+export * from './functions/get-supported-thermostat-fan-modes.function';
+export * from './functions/get-supported-thermostat-modes.function';
+export * from './functions/get-thermostat-fan-mode.function';
+export * from './functions/get-thermostat-mode.function';
+export * from './functions/get-thermostat-operating-state.function';
+export * from './functions/get-thermostat-schedule.function';
+export * from './functions/get-thermostat-setpoint.function';
 
 export * from './functions/set-alarm-setting.function';
+export * from './functions/set-cooling-setpoint.function';
+export * from './functions/set-heating-setpoint.function';
+export * from './functions/set-supported-thermostat-fan-modes.function';
+export * from './functions/set-supported-thermostat-modes.function';
+export * from './functions/set-thermostat-fan-mode.function';
+export * from './functions/set-thermostat-mode.function';
+export * from './functions/set-thermostat-operating-state.function';
+export * from './functions/set-thermostat-schedule.function';
+export * from './functions/set-thermostat-setpoint.function';
 
 /* TYPES */
 
 export * from './types/alarm-setting.type';
+export * from './types/thermostat-fan-modes.type';
+export * from './types/thermostat-modes.type';
+export * from './types/thermostat-operating-modes.type';
