+++ src/hubitat-capabilities/types/thermostat-fan-modes.type.ts
@@ -1,0 +1,9 @@
+/**
+ * @packageDocumentation
+ * @module HubitatCapabilities
+ */
+
+/**
+ * Possible thermostat fan modes.
+ */
+export type EThermostatFanModes = 'on' | 'circulate' | 'auto';
