+++ src/hubitat-capabilities/types/thermostat-modes.type.ts
@@ -1,0 +1,9 @@
+/**
+ * @packageDocumentation
+ * @module HubitatCapabilities
+ */
+
+/**
+ * Possible thermostat modes.
+ */
+export type EThermostatModes = 'auto' | 'off' | 'heat' | 'emergency heat' | 'cool';
