+++ src/hubitat-capabilities/types/thermostat-operating-modes.type.ts
@@ -1,0 +1,16 @@
+/**
+ * @packageDocumentation
+ * @module HubitatCapabilities
+ */
+
+/**
+ * Possible thermostat operating modes.
+ */
+export type EThermostatOperatingState =
+  | 'heating'
+  | 'pending cool'
+  | 'pending heat'
+  | 'vent economizer'
+  | 'idle'
+  | 'cooling'
+  | 'fan only';
